chore: move UIComposer components into separate files
