+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedLoaderPresentationAdapter.swift
@@ -1,0 +1,30 @@
+//
+//  FeedLoaderPresentationAdapter.swift
+//  EssentialFeediOS
+//
+//  Created by Nikhil Menon on 2/13/24.
+//
+
+import EssentialFeed
+
+final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
+    private let feedLoader: FeedLoader
+    var presenter: FeedPresenter?
+    
+    init(feedLoader: FeedLoader) {
+        self.feedLoader = feedLoader
+    }
+    
+    func didRequestFeedRefresh() {
+        presenter?.didStartLoadingFeed()
+        feedLoader.load { [weak self] result in
+            switch result {
+            case let .success(feed):
+                self?.presenter?.didFinishLoadingFeed(with: feed)
+            case let .failure(error):
+                self?.presenter?.didFinishLoadingFeed(with: error)
+            }
+            
+        }
+    }
+}
